+++ app/src/main/java/com/daskiworks/ghwatch/NewVersionInfoDialogFragment.java
@@ -39,7 +39,7 @@ public class NewVersionInfoDialogFragment extends DialogFragment {
 
   private static final String TAG = NewVersionInfoDialogFragment.class.getSimpleName();
 
-  private static final String VERSION_VALUE = "1.18";
+  private static final String VERSION_VALUE = "1.19";
 
   protected View view;